refactor: day 10 solution use custom YXPair type for readability
Updated Day10.kt in the Advent of Code Kotlin project to utilize custom typealias YXPair for pair of integers. This encapsulates the role of these pairs as coordinates in the 2D array used to represent the puzzle. This improves code readability, particularly in the context of coordinate transformations and discrete movements in the array.